requirements.txt
argparse==1.2.1
dj-database-url==0.3.0
Django==1.8.3
django-activelink==0.4
django-admin-bootstrapped==2.5.4
django-admin-tools==0.7.2
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-compressor==1.5
django-debug-toolbar==1.3.2
django-mail-templated==0.2.2
django-mailer==1.2.1
django-threadlocals==0.8
django-tinymce==2.0.4
django-widget-tweaks==1.4.1
ecdsa==0.13
Fabric==1.11.1
FirePython==0.9.0
funcsigs==1.0.2
html5lib==0.999999
jsonpickle==0.9.3
lockfile==0.12.2
Mako==1.0.4
MarkupSafe==0.23
mock==1.3.0
MySQL-python==1.2.5
paramiko==1.17.1
pbr==1.10.0
Pillow==2.9.0
#pip==1.5.6
pisa==3.0.33
psycopg2==2.5
pycrypto==2.6.1
reportlab==2.7
setuptools==3.6
six==1.10.0
sqlparse==0.1.19
wsgiref==0.1.2
django-cors-headers==1.1.0
uwsgi

gunicorn
whitenoise

celery==3.1.23
redis==2.10.5